SSE/windsurf-project/requirements.txt
flask==3.0.3
gunicorn==21.2.0
PyJWT==2.9.0
